Refactor and improve the application
This commit refactors the application by splitting the code into multiple files, which improves code organization and reusability. It also adds several improvements to the application, including:

- Barcode detection using the pyzbar library.
- Improved ZPL optimization logic.
- Caching for the easyocr model to improve performance.
- A .gitignore file to exclude unnecessary files from the repository.
