+++ constants.py
@@ -1,0 +1,1 @@
+LABELARY_API_URL = "http://api.labelary.com/v1/printers"
